Resolve rollup complaints about declaration file by not having a declaration file
